Core Data: Replace protocol extension with a method for replacing container.
The protocol extension trick was intended to let tests replace the default core data stack with its own.
This turns out to not work in practice as the test need to launch the containing app, which has code that initializes
the CoreDataManager before BaseTest can do its thing.

This removes the protocol extensions and replaces it with a method for replacing the NSPersistentContainer.