Refactor: A large refactor to finalize removing extraneous state from api.
Sessions are now site centric rather than a combination of account and site.
There was no real need to manage the account separately since sites can have only a single account.

A session dispatcher is now shared to all stores and used by apis.

The session dispatcher and all stores are recreated when a session changes. This ensures that inflight
api requests are not handled by a store belonging to a different session and a different site.

Stores have a currentSiteID property or a currentAccountID property to ensure they save to the correct site or
account. Initializing a store with the relevant ID was chosen over having the stores query the SessionManager as
a way to encapsulate state within stores and aid in testing. This might change to pass a session object that
defines the dispatcher, ID, etc. rather than setting these individually in the initializer.

PostLists instances can be retrived by their filter now. Its assumed filters will be unique between sites.

SessionManager now listens for deleted accounts and attempts to reinitialize the session with a new account
and new site if found.

Tests still need updating.