Data: Refactors how the core data stack is constructed for tests.
With these changes, tests are not responsible for rebuilding their own core data stack.
The core data manager will check if tests are running and, if so, create an in-memory stack.
A convenience method is created so the stack can be reset between tests.
The managedobjectmodel is reused to avoid warnings about ambiguous entities due to multiple model objects being loaded into memory.
Potentially no longer necessary after iOS 13 -- its supposed to be properly cached then. Will investigate and address if so.